Renames dir
